Bot - process/work api+bot add Router
